feat(components): использование версии проекта из обвязки
Jira issue: VEGA-3336
