+++ src/redux-store/project-structure/actions.ts
@@ -3,10 +3,40 @@ import { TargetData, TreeItem } from '@gpn-prototypes/vega-ui';
 import { AnyAction } from 'redux';
 import { ThunkAction } from 'redux-thunk';
 
+import { clearStores } from '../clear/actions';
+import { setNotification } from '../notifications/actions';
+
 import { ProjectStructureActionTypes } from './action-types';
 
-import { StoreLC } from '@/types/redux-store';
+import { ProjectStructureQuery, StoreLC } from '@/types/redux-store';
 import { logicConstructorService } from '@/utils/lc-service';
+
+interface DomainObject {
+  vid: string;
+  name: string;
+  // eslint-disable-next-line no-underscore-dangle
+  typename: string;
+
+  [key: string]: DomainObject[] | string;
+}
+
+interface Entity {
+  name: string;
+  vid: string;
+}
+
+interface EntityImageAttribute {
+  name: string;
+  attrType: string;
+  entity: Entity;
+}
+
+interface EntityImage {
+  vid: string;
+  name: string;
+  attributes: EntityImageAttribute[];
+  entity: Entity;
+}
 
 export type SetProjectStructureList = {
   type: typeof ProjectStructureActionTypes.SET_PROJECT_STRUCTURE_LIST;
@@ -30,6 +60,11 @@ const setProjectStructureDraggingElements = (
   draggingElements,
 });
 
+const setProjectStructureQuery = (projectStructureQuery: ProjectStructureQuery) => ({
+  type: ProjectStructureActionTypes.SET_PROJECT_STRUCTURE_QUERY,
+  projectStructureQuery,
+});
+
 const DEFAULT_QUERY = `{
   project {
     vid
@@ -42,14 +77,6 @@ const DEFAULT_QUERY = `{
 }`;
 
 const DEFAULT_TREE = ['geoEconomicAppraisalProjectList'];
-
-interface DomainObject {
-  vid: string;
-  name: string;
-  typename: string;
-
-  [key: string]: DomainObject[] | string;
-}
 
 // TODO: make maps with entities
 const ICONS_MAP: { [key: string]: string } = {
@@ -81,12 +108,14 @@ const fetchProjectStructureList = (): ThunkAction<void, StoreLC, unknown, AnyAct
     const query = gql(state.projectStructure.projectStructureQuery?.query || DEFAULT_QUERY);
     const tree = state.projectStructure.projectStructureQuery?.tree || DEFAULT_TREE;
 
-    const response = await logicConstructorService.client?.query({
-      query,
-      context: {
-        uri: logicConstructorService.getGraphQlUri(),
-      },
-    });
+    const response = await logicConstructorService.client
+      ?.watchQuery({
+        query,
+        context: {
+          uri: logicConstructorService.getGraphQlUri(),
+        },
+      })
+      .result();
 
     if (response?.data) {
       const { domain } = response.data.project;
@@ -104,4 +133,77 @@ const fetchProjectStructureList = (): ThunkAction<void, StoreLC, unknown, AnyAct
   }
 };
 
-export { fetchProjectStructureList, setProjectStructureList, setProjectStructureDraggingElements };
+function buildStructureQuery(entityImages: EntityImage[]): ProjectStructureQuery {
+  if (!entityImages.length) return { query: '', tree: [] };
+
+  const image = entityImages.find((ei) => ei.name === 'GeoEconomicAppraisalProject');
+  let query = `{ project { vid domain { geoEconomicAppraisalProjectList { typename:__typename vid name `;
+  const tree = ['geoEconomicAppraisalProjectList'];
+  const loadedImages: string[] = [];
+
+  function buildAttributeQuery(attr: EntityImageAttribute | undefined): string {
+    if (!attr) {
+      return '';
+    }
+    const found = entityImages.find(
+      (ei) => ei.entity.vid === attr.entity.vid && loadedImages.indexOf(ei.vid) < 0,
+    );
+    if (found) {
+      tree.push(attr.name);
+      loadedImages.push(found.vid);
+      const attributeFound = found.attributes.find((i) => i.attrType === '[*]');
+
+      return `${attr.name} { typename:__typename ... on ${
+        found.name
+      }_Type { vid name ${buildAttributeQuery(attributeFound)} } }`;
+    }
+
+    return '';
+  }
+
+  if (image) {
+    const attributeFound = image.attributes.find((i) => i.attrType === '[*]');
+    query += buildAttributeQuery(attributeFound);
+  }
+
+  query += '} } } }';
+
+  return {
+    query,
+    tree,
+  };
+}
+
+const fetchProjectSchema = (): ThunkAction<void, StoreLC, unknown, AnyAction> => async (
+  dispatch,
+): Promise<void> => {
+  dispatch(clearStores());
+
+  try {
+    const response = await logicConstructorService.projectStructureQuery();
+
+    if (response?.data) {
+      const structureQuery = buildStructureQuery(response.data?.project.domainSchema.entityImages);
+
+      dispatch(setProjectStructureQuery(structureQuery));
+    } else {
+      dispatch(setNotification({ message: 'Пустой ответ сервера', status: 'alert' }));
+    }
+  } catch (e) {
+    const { errors } = e.networkError?.result;
+    if (
+      Array.isArray(errors) &&
+      errors.find((error) => error.message === 'badly formed hexadecimal UUID string')
+    ) {
+      const message = 'В url не корректный UUID проекта';
+      dispatch(setNotification({ message, status: 'alert' }));
+    }
+  }
+};
+
+export {
+  fetchProjectSchema,
+  fetchProjectStructureList,
+  setProjectStructureList,
+  setProjectStructureDraggingElements,
+};
