--- src/redux-store/version/initial-state.ts
@@ -1,7 +1,0 @@
-import { VersionState } from '@/types/redux-store';
-
-const initialState: VersionState = {
-  version: -1,
-};
-
-export default initialState;
